+++ src/renderer/src/App.tsx
@@ -48,6 +48,7 @@ import { UpdateModal } from '#/renderer/src/ui/modals/UpdateModal';
 import { AlertModal } from '#/renderer/src/ui/modals/AlertModal';
 import { CollectionModal } from '#/renderer/src/ui/modals/CollectionModal';
 import { ConfirmModal } from '#/renderer/src/ui/modals/ConfirmModal';
+import { PluginThemePrompt } from '#/renderer/src/ui/modals/PluginThemePrompt';
 import { ShareModal } from '#/renderer/src/ui/modals/ShareModal';
 import { CollectionRunnerModal } from '#/renderer/src/ui/modals/CollectionRunnerModal';
 import { QuitPrompt } from '#/renderer/src/ui/modals/QuitPrompt';
@@ -303,6 +304,7 @@ export default function App(): JSX.Element {
           <CollectionRunnerModal />
           <AlertModal />
           <ConfirmModal />
+          <PluginThemePrompt />
 
           <Toaster
             position="bottom-center"